Rename core modules for clarity
