+++ game_core/debug_manager.py
