+++ game_core/edit_mode/edit_screen_ui.py
@@ -2,7 +2,7 @@
 Edit Screen UI module - handles drawing the UI elements for the edit screen
 """
 import pygame
-from settings import *
+from game_settings import *
 from font_manager import font_manager
 
 class EditScreenUI:
