+++ game_core/edit_mode/relations.py
@@ -2,7 +2,7 @@
 Relation components for the Edit Mode
 """
 import pygame
-from settings import *
+from game_settings import *
 from font_manager import font_manager
