--- game_core/enemy_system/__init__.py
@@ -1,8 +1,0 @@
-"""
-Enemy System Package - Contains classes for managing game enemies
-"""
-from enemy_system.enemy import Enemy
-from enemy_system.phantom import Phantom
-from enemy_system.bomberplant import Bomberplant
-from enemy_system.spinner import Spinner
-from enemy_system.enemy_manager import EnemyManager
