+++ game_core/playscreen_components/enemy_system/enemy_manager.py
@@ -2,13 +2,13 @@
 Enemy Manager - Manages all enemies in the game
 """
 import pygame
-from enemy_system.phantom import Phantom
-from enemy_system.bomberplant import Bomberplant
-from enemy_system.spinner import Spinner
-from enemy_system.spider import Spider
-from enemy_system.pinkslime import Pinkslime
-from enemy_system.pinkbat import Pinkbat
-from debug_utils import debug_manager
+from .phantom import Phantom
+from .bomberplant import Bomberplant
+from .spinner import Spinner
+from .spider import Spider
+from .pinkslime import Pinkslime
+from .pinkbat import Pinkbat
+from debug_manager import debug_manager
 from performance_optimizer import performance_optimizer
 
 class EnemyManager:
