+++ game_core/playscreen_components/enemy_system/phantom.py
@@ -5,8 +5,8 @@ import pygame
 import os
 import heapq
 import math
-from enemy_system.enemy import Enemy
-from debug_utils import debug_manager
+from .enemy import Enemy
+from debug_manager import debug_manager
 
 class Phantom(Enemy):
     """Phantom enemy class with cardinal movement"""
@@ -299,7 +299,7 @@ class Phantom(Enemy):
 
             # Debug output
             if self.debug_mode:
-                from debug_utils import debug_manager
+                from debug_manager import debug_manager
                 debug_manager.log(f"Applied knockback to player: dx={dx:.2f}, dy={dy:.2f}", "enemy")
 
             return True
