+++ game_core/playscreen_components/inventory_system/chest_inventory.py
@@ -3,7 +3,7 @@ Chest Inventory - handles the display and interaction with lootchest contents (T
 """
 import pygame
 import os
-from game_core.sprite_cache import sprite_cache
+from game_core.sprite_cache_system import sprite_cache
 
 class ChestInventory:
     """Class to handle the lootchest inventory display with Terraria-style interactions"""
