+++ game_core/playscreen_components/player_system/sprite_loader.py
@@ -3,7 +3,7 @@ Sprite loader module - handles loading character sprites and animations
 """
 import os
 import pygame
-from game_core.sprite_cache import sprite_cache
+from game_core.sprite_cache_system import sprite_cache
 
 def load_character_sprites():
     """Load all character sprites from the character folder"""
