+++ game_core/playscreen_components/ui_system/hud.py
@@ -1,7 +1,7 @@
 import pygame
 import os
 import math
-from game_core.sprite_cache import sprite_cache
+from game_core.sprite_cache_system import sprite_cache
 
 class Inventory:
     """Class to handle the player's inventory"""
