+++ game_core/sprite_cache_system.py
